Explain GHC that vector's indices are strict
This is to prevent degradation when vector's flavour does not get
specialized early enough, in which case GHC may decide that
(!), (!.~) and (<~~) do not necessarily use indices strictly.
This has catastrophic consequences for performance, even beyond
usual penalties for unspecialized vectors, because indices
will passed as boxed Ints.

Hopefully this will be eventually fixed in the vector package itself
(https://github.com/haskell/vector/pull/485), but for now it does not
hurt to be explicit on the side of vector-hashtables.